BeadRec: fix scan-arming race, add live ROI preview and orientation controls
- Scan-arming race ("shifted/extra frames" at scan start): BeadWorker used to
  gate frame consumption on raw isScanRunning, which flips True while scan
  signals are still being armed, before any real scan frame exists -- pulling
  the pre-scan camera backlog into the reconstruction and shifting/inflating
  it. Now gated on _scanFramesReady (scanArmed AND isScanRunning), with
  _scanArmed only set after the pre-scan buffer is explicitly flushed
  (_discardBufferedFrames) in onNewScan. Added a frame-count diagnostic in
  onEndedScan comparing received vs expected (dims) frames.
- Live ROI crop preview: while the ROI is shown and not reconstructing, mirror
  exactly what the ROI crops from the current live detector frame into the
  BeadRec viewer, so the user can see what each reconstruction pixel will
  average from before starting a scan. Reuses the same display-transform path
  as the reconstruction crop (see below) so the preview matches what's drawn.
- Reconstruction display-transform / ROI-region fix: the ROI is drawn on the
  DISPLAYED (rotated/flipped per detector displayRotation) image, but
  readChunk()/live frames are raw. Added _currentDisplayTransform /
  _toDisplayedFrames so the chunk getter and the live preview both crop the
  same transformed frame the ROI was drawn against -- otherwise cameras with a
  non-identity display orientation reconstruct from a shifted region.
- Orientation controls (rotate 90 deg steps + flip H/V) for the finished
  reconstruction, to correct for the physical scan direction/origin corner;
  self._orientBase keeps the pre-orientation image so toggling re-renders
  without recomputing the scan.
- bead_recognition.mean_intensity_in_roi: force a float64 mean accumulator
  explicitly, since frames are now native integer dtype (e.g. uint16) rather
  than the float64 BeadRec used to receive -- guards against a truncated
  (non-fractional) ROI mean.
- New utility_scripts/beadrec_reconstruct_offline.py: reproduces the live
  BeadRec reconstruction (whole-frame mean per pixel, laid out via scan-grid
  metadata) from an already-recorded HDF5/Zarr stack, for offline
  re-reconstruction without rerunning the scan.

Tests: test_bead_rec_controller_contract.py, test_bead_recognition_model.py
(updated/added cases for the new gating callable and float-mean guard).

+++ imswitch/imcontrol/model/bead_recognition.py
@@ -236,7 +236,12 @@ def mean_intensity_in_roi(frame: np.ndarray, roi: RoiBounds) -> float:
         raise ValueError("Bead reconstruction expects 2D detector frames")
     clipped_roi = normalize_roi_bounds((roi.x0, roi.y0, roi.x1, roi.y1), frame.shape)
     rows, cols = clipped_roi.as_slices()
-    return float(np.mean(frame[rows, cols]))
+    # Force a float64 accumulator. np.mean already promotes integer inputs to
+    # float64, but the detector dtype contract now hands BeadRec native integer
+    # frames (e.g. uint16) rather than the float64 it used to receive, so make
+    # the float intent explicit: a fractional ROI mean is never truncated, and
+    # float32 frames don't lose precision when averaging over a large ROI.
+    return float(np.mean(frame[rows, cols], dtype=np.float64))
 
 
 def create_reconstruction_buffer(scan_dims: Sequence[int]) -> np.ndarray:
